cache_server/requirements.txt
grpcio==1.44.0
protobuf==3.19.4
requests==2.26.0
